refactor(rvr): clarify RV64 extension ownership
