+++ extensions/riscv/circuit/src/extension/mod.rs
@@ -34,7 +34,8 @@ use openvm_stark_backend::p3_field::PrimeField32;
 use openvm_stark_backend::{StarkEngine, StarkProtocolConfig, Val};
 #[cfg(feature = "rvr")]
 use rvr_openvm_ext_riscv::{
-    Rv64IExtension, Rv64IRuntimeHooks, Rv64IoExtension, Rv64IoRuntimeHooks, Rv64MExtension,
+    Rv64IExtension, Rv64IoExtension, Rv64IoRuntimeHooks, Rv64MExtension, Rv64PhantomExtension,
+    Rv64PhantomRuntimeHooks,
 };
 #[cfg(feature = "rvr")]
 use rvr_openvm_lift::{RvrExtensionCtx, RvrExtensions, VmRvrExtension};
@@ -94,7 +95,8 @@ fn default_range_tuple_checker_sizes() -> [u32; 2] {
 impl<F: PrimeField32> VmRvrExtension<F> for Rv64I {
     fn extend_rvr(&self, extensions: &mut RvrExtensions, _ctx: Option<&RvrExtensionCtx>) {
         extensions.register_lifter(Rv64IExtension::new());
-        extensions.register_runtime_hook(Rv64IRuntimeHooks);
+        extensions.register_lifter(Rv64PhantomExtension::new());
+        extensions.register_runtime_hook(Rv64PhantomRuntimeHooks);
     }
 }
