--- extensions/riscv/rvr/c/rv64i_phantom_callbacks.h
@@ -1,23 +1,0 @@
-#ifndef RV64I_PHANTOM_CALLBACKS_H
-#define RV64I_PHANTOM_CALLBACKS_H
-
-#include <stdint.h>
-
-typedef struct {
-  void (*hint_input)(void* ctx);
-  bool (*print_str)(void* ctx, uint64_t ptr, uint64_t len);
-  bool (*hint_random)(void* ctx, uint64_t num_words);
-} Rv64IHostCallbacks;
-
-/* Status-returning callbacks and stubs return false for invalid guest operands. */
-
-void register_rv64i_host_callbacks(const Rv64IHostCallbacks* cb);
-
-/* Forwarding stubs owned by the RISC-V base (Rv64I) extension: the IO phantoms
- * (HINT_INPUT, PRINT_STR, HINT_RANDOM). Backed by a thread-local dispatch table
- * installed at execution time by `Rv64IRuntimeHooks`. */
-void openvm_hint_input(void);
-bool openvm_print_str(uint64_t ptr, uint64_t len);
-bool openvm_hint_random(uint64_t num_words);
-
-#endif /* RV64I_PHANTOM_CALLBACKS_H */
