+++ extensions/riscv/rvr/c/rv64m.h
@@ -64,7 +64,7 @@ static __attribute__((always_inline)) inline uint64_t rv_mulhsu(int64_t a,
   return (uint64_t)(((__int128)a * (__int128)b) >> 64);
 }
 
-/* ── W-suffix helpers (rv64): operate on low 32 bits, sign-extend to 64 ── */
+/* ── W-suffix helpers (RV64): operate on low 32 bits, sign-extend to 64 ── */
 
 static __attribute__((always_inline)) inline uint64_t rv_divw(int32_t a,
                                                               int32_t b) {
